refactor: Profile View 분리
